refactor: #105 extract normalize() & createCriteriaForIndex() methods
